# docs/swagger.yaml
definitions:
  aggregate.DeveloperTask:
    properties:
      developer:
        $ref: '#/definitions/entity.Developer'
      tasks:
        items:
          $ref: '#/definitions/valueobject.Task'
        type: array
      week:
        type: integer
    type: object
  dto.AddProviderRequest:
    properties:
      name:
        type: string
      task_duration_key:
        type: string
      task_name_key:
        type: string
      task_value_key:
        type: string
      url:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: boolean
      msg:
        type: string
    type: object
  dto.Provider:
    properties:
      id:
        type: string
      name:
        type: string
      task_duration_key:
        type: string
      task_name_key:
        type: string
      task_value_key:
        type: string
      tasks:
        items:
          $ref: '#/definitions/valueobject.Task'
        type: array
      url:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  dto.Task:
    properties:
      difficulty:
        type: integer
      duration:
        type: integer
      id:
        type: string
    type: object
  entity.Developer:
    properties:
      id:
        type: string
      name:
        type: string
      remainingHours:
        type: integer
      speed:
        type: integer
    type: object
  valueobject.Task:
    properties:
      difficulty:
        type: integer
      duration:
        type: integer
      id:
        type: string
      intensity:
        type: integer
      name:
        type: string
      providerID:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/provider:
    post:
      consumes:
      - application/json
      description: Add a new provider to add tasks
      operationId: insert-provider
      parameters:
      - description: Add provider request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Add a new provider
  /v1/providers:
    get:
      description: Get all the task from the database
      operationId: get-tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all the tasks
  /v1/task:
    post:
      description: Assign tasks to developers
      operationId: assign-tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/aggregate.DeveloperTask'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Assign tasks to developers
swagger: "2.0"